refactor: minor views and templates restructuring
